requirements.txt
streamlit==1.17.0
google-generativeai==1.0.0  # Make sure to check and use the correct version for your library
requests==2.26.0
beautifulsoup4==4.10.0
python-dotenv==0.19.2